pyproject.toml
[project]
name = "worldcup-2026-prediction"
version = "0.1.0"
description = "End-to-end pipeline for World Cup 2026 and qualifiers match prediction."
requires-python = ">=3.12"
dependencies = [
  "SQLAlchemy==2.0.48",
  "dbt-core==1.11.7",
  "dbt-postgres==1.10.0",
  "fastapi==0.135.3",
  "gunicorn>=25.3.0",
  "httptools>=0.7.1",
  "ipykernel==7.2.0",
  "joblib==1.5.3",
  "matplotlib==3.10.8",
  "numpy==2.4.4",
  "pandas==3.0.2",
  "plotly>=6.7.0",
  "psycopg2-binary==2.9.11",
  "python-dotenv==1.2.2",
  "requests==2.33.1",
  "scikit-learn==1.8.0",
  "seaborn==0.13.2",
  "streamlit>=1.56.0",
  "uvicorn==0.42.0",
  "xgboost==3.2.0",
]

[project.urls]
"Homepage" = "https://worldcup-2026-prediction.onrender.com"
"Repository" = "https://github.com/JuanjoRestrepo/worldcup-2026-prediction"
"Documentation" = "https://worldcup-2026-prediction.onrender.com/docs"

[dependency-groups]
dev = [
  "httpx==0.28.1",
  "jupyter>=1.1.1",
  "jupytext>=1.19.1",
  "mypy==1.18.2",
  "nbconvert>=7.17.1",
  "pandas-stubs==2.3.2.250926",
  "pytest==9.0.2",
  "ruff>=0.11,<0.13",
  "shap>=0.51.0",
]

[tool.uv]
package = false
default-groups = ["dev"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings
  "F",    # pyflakes
  "I",    # isort
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "UP",   # pyupgrade
  "N",    # pep8-naming
  "S",    # flake8-bandit (security)
  "PTH",  # use pathlib over os.path
]
ignore = [
  "E501",   # line too long — formatter handles wrapping; long strings in Field() descriptions are acceptable
  "S608",   # possible SQL injection
  "E402",   # module level import not at top of file
  "N803",   # argument name should be lowercase (conflicts with X in sklearn)
  "N806",   # variable in function should be lowercase (conflicts with X in sklearn)
  "S101",   # assert used in tests — pytest relies on assert
  "S105",   # hardcoded password — false positives on env var defaults
  "S106",   # hardcoded password in function call
  "B008",   # function calls in default arguments — FastAPI Depends pattern
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S", "N802", "N803"]   # security + naming conventions relaxed in tests
"run_pipeline.py" = ["T201"]        # print() allowed in pipeline runner
"run_dbt.py" = ["T201"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
files = ["src/api", "src/modeling", "src/database", "src/contracts"]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
pretty = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["joblib", "psycopg2.*", "sklearn.*", "xgboost.*", "dbt.*"]
ignore_missing_imports = true
